refactor: error 패키지 구조 이동
