+++ src/main/java/roomescape/common/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.common.exception;
 
 public class BusinessException extends RuntimeException {
