+++ src/main/java/roomescape/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.common.exception;
 
 import org.springframework.http.HttpStatus;
