+++ src/main/java/roomescape/common/exception/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package roomescape.exception.dto;
+package roomescape.common.exception.dto;
 
 import org.springframework.http.HttpStatus;
