+++ src/test/java/roomescape/unit/ReservationServiceTest.java
@@ -13,7 +13,7 @@ import roomescape.domain.reservation.dto.ReservationResponse;
 import roomescape.domain.reservation.dto.ReservationUpdateRequest;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
-import roomescape.exception.BusinessException;
+import roomescape.common.exception.BusinessException;
 import roomescape.repository.*;
 import roomescape.service.ReservationService;
